movie_pjt/requirements.txt
asgiref==3.4.1
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Django==3.2.11
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2022.1
urllib3==1.26.9
